[jupyter] allow import of python code from the src/python_modules dir of validation reports
goal is to promote better organisation of scripts when complex postprocessing is needed

currently, reports are using a mix of bash and python scripts mixed with a lot of os.chdirs and other operations. this makes debugging report a complete hell. Hopefully a pure python approach may be better in time